Separated layers via folders in src
