+++ src/Logic/PowerCalculator.cpp
