+++ src/Logic/PowerCalculator.h
