+++ src/Logic/Utilities.cpp
