+++ src/SolarSimulator.cpp
@@ -1,12 +1,12 @@
 #include "SolarSimulator.h"
-#include "Window.h"
-#include "FileLoader.h"
-#include "ImageData.h"
-#include "Shader.h"
-#include "Texture.h"
-#include "Mesh.h"
-#include "ModelData.h"
-#include "Utilities.h"
+#include "View/Window.h"
+#include "Logic/FileLoader.h"
+#include "Data/ImageData.h"
+#include "Presenter/Shader.h"
+#include "Presenter/Texture.h"
+#include "Presenter/Mesh.h"
+#include "Data/ModelData.h"
+#include "Logic/Utilities.h"
 #include <GL/glew.h>
 #include <glm/glm.hpp>
 #include <string>
